feat: Serve places at places base path
