+++ .gitignore
@@ -58,6 +58,7 @@ typings/
 .parcel-cache/
 .cache/
 public
+public-prefix
 
 # Mac files
 .DS_Store
