+++ gatsby-config.js
@@ -9,6 +9,7 @@ module.exports = {
     description: `Decentraland`,
     author: `@decentraland`,
   },
+  pathPrefix: "/places",
   developMiddleware: developMiddleware({
     prefix: `/api`,
     url: `http://127.0.0.1:4000`,
