+++ src/config/dev.json
@@ -1,8 +1,9 @@
 {
+  "GATSBY_NEW_ROLLOUT": "true",
   "GATSBY_CHAIN_ID": "1,137",
   "GATSBY_DECENTRALAND_URL": "https://play.decentraland.org",
   "GATSBY_LAND_URL": "https://api.decentraland.org",
-  "GATSBY_PLACES_URL": "/api",
+  "GATSBY_PLACES_URL": "https://places.decentraland.zone/api",
   "GATSBY_PROFILE_URL": "https://peer.decentraland.org",
   "GATSBY_INTERCOM_APP_ID": "z0h94kay",
   "GATSBY_SENTRY_SRC": "https://js.sentry-cdn.com/bb04b48d542c5a48ad9f41ea2429a500.min.js",
